Auth Application Completed

+++ Backend/controllers/goalController.js
@@ -1,11 +1,16 @@
 const asyncHandler = require("express-async-handler");
 const Goal = require("../models/goalsModel");
+const User = require("../models/userModel");
+
 //@desc Get Goals
 //@route GEt /api/goals
 //@access private
+
+
+
 const getGoals = () =>
   asyncHandler(async (req, res) => {
-    const goals = await Goal.find();
+    const goals = await Goal.find({ user: req.user.id });
     res.status(200).json({ goals });
   });
 
@@ -19,7 +24,7 @@ const setGoal = () =>
       res.status(400);
       throw new Error("Please provide a goal");
     }
-    const newGoal = await Goal.create({ goal });
+    const newGoal = await Goal.create({ goal, user: req.user.id });
     return res.status(201).json({ message: "Success", goal: newGoal });
   });
 
@@ -35,6 +40,17 @@ const updateGoal = () =>
       res.status(400);
       throw new Error("Goal not found");
     }
+
+    const user = await User.findById(req.user.id);
+    if (!user) {
+      res.status(401);
+      throw new Error("User not found");
+    }
+    //check if user owns this goalToUpdate
+    if (goalToUpdate.user.toString() !== user.id) {
+      res.status(401);
+      throw new Error("User not authorized");
+    }
     const updatedGoal = await Goal.findByIdAndUpdate(
       id,
       { goal },
@@ -49,11 +65,24 @@ const updateGoal = () =>
 const deleteGoal = () =>
   asyncHandler(async (req, res) => {
     const { id } = req.params;
+
     const goalToUpdate = await Goal.findById(id);
     if (!goalToUpdate) {
       res.status(400);
       throw new Error("Goal not found");
     }
+
+    const user = await User.findById(req.user.id);
+    if (!user) {
+      res.status(401);
+      throw new Error("User not found");
+    }
+    //check if user owns this goalToUpdate
+    if (goalToUpdate.user.toString() !== user.id) {
+      res.status(401);
+      throw new Error("User not authorized");
+    }
+
     await Goal.findByIdAndDelete(id);
 
     res.status(200).json({ message: `Delete Goals ${req.params.id}` });
